+++ .vscode/launch.json
@@ -4,27 +4,11 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [{
-        "name": "example",
+        "name": "ut_kuksa",
         "type": "cppdbg",
         "request": "launch",
-        "program": "${workspaceFolder}/build/Release/example/example",
-        "args": [],
-        "stopAtEntry": true,
-        "cwd": "${fileDirname}",
-        "environment": [],
-        "externalConsole": false,
-        "MIMode": "gdb",
-        "setupCommands": [
-            {
-                "description": "Enable pretty-printing for gdb",
-                "text": "-enable-pretty-printing",
-                "ignoreFailures": true
-            },
-            {
-                "description": "Set Disassembly Flavor to Intel",
-                "text": "-gdb-set disassembly-flavor intel",
-                "ignoreFailures": true
-            }
-        ]
+        "cwd": ".",
+        "program": "${workspaceFolder}/build/Release/bin/ut_kuksa",
+        "stopAtEntry": true
     }]
 }
